fix : index로 수정
